requirements.txt
# Discord Integration
discord.py>=2.3.0

# Data Validation
pydantic>=2.0.0

# Configuration Management
python-dotenv>=1.0.0

# YAML Support
PyYAML>=6.0

# AI Integration
openai>=1.0.0

# Database
aiosqlite>=0.19.0
libsql-client>=0.3.0

# SSL/TLS Certificates
certifi>=2023.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type Checking (Development)
mypy>=1.5.0

# Code Quality (Development)
black>=25.9.0
ruff>=0.0.285
pre-commit>=4.3.0
